Switch to rv32i toolchain
Originally, I used an 13.2.0 rv64 toolchain with multilib support
without issues. However, upon upgrading to gcc version 15.1.0,
tests were failing. This was due to compressed nops being inserted into
the binary, despite compiling the toolchain with either multilib
support, which supports rv32i, or compiling the toolchain exclusively
for rv32i_zicsr. I traced the issue back to a problem with linker
relaxation and the rvc + norvc options, as described in the following
issues:

- https://github.com/riscv-non-isa/riscv-asm-manual/issues/84
- https://github.com/riscv-collab/riscv-gnu-toolchain/issues/445

For whatever reason, when nops are inserted for alignment and the
rvc + norvc options are used, the linker will relax certain nops to
compressed nops, which my core doesn't support.

The solution was to prevent any linker relaxation with the -mno-relax
option.